pyproject.toml
[project]
name = "merchantbot"
version = "0.1.0"
description = "MCMerchant Discord support bot built with discord.py"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.5.2",
  "aiohttp>=3.11.18",
  "aiosqlite>=0.20.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.8.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "pytz>=2024.1",
  "cachetools>=5.5.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["merchantbot"]

